+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Improves and simplifies tests
 * Fixes code to work on cropped stars proxy
+* Adds a type argument to `lsp_add_quality()`, making it possible to calculate either isolation (type = "segmentation") or distinction (type = "cluster")
 
 # motif 0.3.12